Finish project content page
I finished building all the functions for the project
specific content pages. As of now you can create
Todos, edit and delete them. I plan on writing functions
to sort the Todos by priority, name and due date.

Aside from that I still need to write the functions
for today's, this week's and this month's tasks.