Build functions to sort todos and show all todos
I built functionality to sort the todos of a project
by name, due date or priority. I also began building
the functions to display all and today's todos. As
of now changes made in the "all todos" and "today's
todos" are not reflected within their corresponding
project. I'll need to fix this and also add this
week's and month's todo display before I tackle CSS
styles to finish the project.